+++ src/bitsea/downloader/old_havesters/biofloats_harvester.py
@@ -42,7 +42,7 @@ class BioFloatsHarvester(HarvesterInterface):
     in a xml file whose path is set in the global variable "xml_path".
     """
     def wmo_file_reader(self):
-        FLOAT_dtype=[('id',np.int),('wmo','S20'),('id_type',np.int),('type','S20'),('nome_fs','S20'),('status','S1')]
+        FLOAT_dtype=[('id',int),('wmo','S20'),('id_type',int),('type','S20'),('nome_fs','S20'),('status','S1')]
         TABLE=np.loadtxt(wmo_file,dtype=FLOAT_dtype,skiprows=2,delimiter=" | ")
         return TABLE
